requirements.txt
google-genai==1.56.0
langchain==1.0.4
langgraph==1.0.2
numpy==2.2.6
pandas==2.3.3
python-dotenv==1.2.1
streamlit==1.53.0
supabase==2.27.2